requirements.txt
PyYAML==3.13
PyInstaller==3.4